Add prism framework and restructure app
